+++ cmd2.py
@@ -113,13 +113,17 @@ try:
 except ImportError:
     ipython_available = False
 
-# Try to import readline, but allow failure for convenience in Windows unit testing
-# Note: If this actually fails, you should install readline on Linux or Mac or pyreadline on Windows
+# Prefer statically linked gnureadline if available (for macOS compatibility due to issues with libedit)
 try:
-    # noinspection PyUnresolvedReferences
-    import readline
+    import gnureadline as readline
 except ImportError:
-    pass
+    # Try to import readline, but allow failure for convenience in Windows unit testing
+    # Note: If this actually fails, you should install readline on Linux or Mac or pyreadline on Windows
+    try:
+        # noinspection PyUnresolvedReferences
+        import readline
+    except ImportError:
+        pass
 
 # Load the GNU readline lib so we can make changes to it
 readline_lib = None